+++ src/Presentation/SmartStore.Web.Framework/WebApi/WebApiCore.cs
@@ -10,10 +10,11 @@ namespace SmartStore.Web.Framework.WebApi
 		public static int DefaultTimePeriodMinutes { get { return 15; } }
 		public static string RouteNameDefaultApi { get { return "WebApi.Default"; } }
 		public static string RouteNameDefaultOdata { get { return "WebApi.OData.Default"; } }
+		public static string RouteNameUploads { get { return "WebApi.Uploads"; } }
 		public static string MostRecentOdataPath { get { return "odata/v1"; } }
 		public static string PluginSystemName { get { return "SmartStore.WebApi"; } }
 
-		/// <remarks>see http://tools.ietf.org/html/rfc6648</remarks>
+		/// <see cref="http://tools.ietf.org/html/rfc6648"/>
 		public static class Header
 		{
 			private static string Prefix { get { return "SmartStore-Net-Api-"; } }