+++ src/Plugins/SmartStore.WebApi/changelog.md
@@ -1,5 +1,9 @@
 ﻿#Release Notes
 
+##Web Api 2.2.0.5
+###New Features
+* Bridge to import framework: uploading import files to import profile directory
+
 ##Web Api 2.2.0.4
 ###New Features
 * Added OData endpoint for shipment items